third_party/WebKit/LayoutTests/accessibility/aria-none-role-expected.txt

Link and text

This tests that the aria 'none' role works by successfully removing the element from the AX tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS firstChild.role is 'AXRole: AXLink'
PASS successfullyParsed is true

TEST COMPLETE
